Fixed overflow bug in vobsub reading :)
git-svn-id: https://subtitleedit.googlecode.com/svn/trunk@109 99eadd0c-20b8-1223-b5c4-2a2b2df33de2

+++ src/Logic/VobSub/PacketizedElementaryStream.cs
@@ -67,7 +67,7 @@ namespace Nikse.SubtitleEdit.Logic.VobSub
             {
                 string bString = Helper.GetBinaryString(buffer, tempIndex, 5);
                 bString = bString.Substring(4, 3) + bString.Substring(8, 15) + bString.Substring(24, 15);
-                PresentationTimeStamp = Convert.ToUInt32(bString, 2);
+                PresentationTimeStamp = Convert.ToUInt64(bString, 2);
                 tempIndex += 5;
             }
             if (PresentationTimeStampDecodeTimeStampFlags == Helper.B00000011)
